Moved getMergedConfig function to separate service
