+++ src/utils/node/tailwindSafelist.js
@@ -7,7 +7,7 @@ import { isEqual, omit } from "lodash-es";
 import { defineConfig } from "cva";
 
 import { vuelessConfig } from "./vuelessConfig.js";
-import { createMergeConfigsFunction } from "./mergeConfigs.js";
+import { createGetMergedConfigFunction } from "./mergeConfigs.js";
 import { getDefaultConfigJson, getDirFiles } from "./helper.js";
 import {
   COMPONENTS,
@@ -20,6 +20,7 @@ import {
   NESTED_COMPONENT_REG_EXP,
   TAILWIND_COLOR_OPACITY_DELIMITER,
   TAILWIND_VARIANT_DELIMITER_REG_EXP,
+  STRATEGY_TYPE,
 } from "../../constants.js";
 
 const twMerge = extendTailwindMerge(merge(TAILWIND_MERGE_EXTENSION, vuelessConfig.tailwindMerge));
@@ -30,7 +31,7 @@ export const { cx } = defineConfig({
   },
 });
 
-const mergeConfigs = createMergeConfigsFunction(cx);
+const getMergedConfig = createGetMergedConfigFunction(cx);
 
 export function clearTailwindSafelist() {
   process.env.VUELESS_SAFELIST = "";
@@ -174,9 +175,15 @@ async function getComponentSafelist(componentName, { colors, vuelessConfigFiles 
     defaultConfig = getDefaultConfigJson(defaultConfigContent);
   }
 
-  const mergedConfig = omit(mergeConfigs({ defaultConfig, globalConfig: customConfig }), [
-    "defaults",
-  ]);
+  const isStrategyValid =
+    vuelessConfig.strategy && Object.values(STRATEGY_TYPE).includes(vuelessConfig.strategy);
+
+  const vuelessStrategy = isStrategyValid ? vuelessConfig.strategy : STRATEGY_TYPE.merge;
+
+  const mergedConfig = omit(
+    getMergedConfig({ defaultConfig, globalConfig: customConfig, vuelessStrategy }),
+    ["defaults"],
+  );
 
   const colorString = `(${colors.join("|")})`;
